ci/tools: add patch test system
While the DPDK community is increasing in size, it becomes more and more
important to run some CI system that will validate each patch.

In order to support CI, it is necessary given a patch ID to find to
which git repository the patch should be applied to.

This commit address this issue. When running the cit.py file with the
following parameters <starting commit> <number of commits> it logs the
patch base info and the candidates repositories ordered by repository
priority.
The possible repositories and the rules for selecting the repositories
can be found in the git-repoitories.json file.

The system is build ontop of the pwclient, which is used in order to
find and download the patches.

In future rleases this system will also download the patch, apply it to
the correct repository and run the selected tests. The test results will
be published using the current system.

File description:
cit.py - The starting point for this system.
classify.py - Select the repositories that match the criteria for the
selected patch.
git-repositories.json - Holds the rules for each git repository. There
are two kinds of rules. Based on patch subject or based on the files
affected by this patch.
patch.py - Holds one patch information.
patch_list.py - Holds a list of patches.
pwclass.py - A wrapper for the pwclient

Signed-off-by: Ori Kam <orika@mellanox.com>